Remove misnamed duplicate zone folder..
