--- scripts/zones/Escha-ZiTah/TextIDs.lua
@@ -1,8 +1,0 @@
--- Variable TextID   Description text
-
--- General Texts
-   ITEM_CANNOT_BE_OBTAINED = 6379; -- You cannot obtain the item <item>. Come back after sorting your inventory.
-             ITEM_OBTAINED = 6382; -- Obtained: <item>.
-              GIL_OBTAINED = 6383; -- Obtained <number> gil.
-          KEYITEM_OBTAINED = 6385; -- Obtained key item: <keyitem>.
-            ITEMS_OBTAINED = 6388; -- You obtain <param2 number> <param1 item>!
