--- scripts/zones/Escha-ZiTah/Zone.lua
@@ -1,55 +1,0 @@
------------------------------------
---
--- Zone: Escha - Zi'Tah (288)
---
------------------------------------
-package.loaded["scripts/zones/Escha-ZiTah/TextIDs"] = nil;
------------------------------------
-
-require("scripts/zones/Escha-ZiTah/TextIDs");
-
------------------------------------
--- onInitialize
------------------------------------
-
-function onInitialize(zone)
-end;
-
------------------------------------
--- onZoneIn
------------------------------------
-
-function onZoneIn( player, prevZone)
-    local cs = -1;
-
-    if (player:getXPos() == 0 and player:getYPos() == 0 and player:getZPos() == 0) then
-        -- player:setPos(x, y, z, rot);
-    end
-
-    return cs;
-end;
-
------------------------------------
--- onRegionEnter
------------------------------------
-
-function onRegionEnter( player, region)
-end;
-
------------------------------------
--- onEventUpdate
------------------------------------
-
-function onEventUpdate( player, csid, option)
-    -- printf("CSID: %u",csid);
-    -- printf("RESULT: %u",option);
-end;
-
------------------------------------
--- onEventFinish
------------------------------------
-
-function onEventFinish( player, csid, option)
-    -- printf("CSID: %u",csid);
-    -- printf("RESULT: %u",option);
-end;
