+++ scripts/zones/Escha_ZiTah/Zone.lua
@@ -1,14 +1,14 @@
 -----------------------------------
 --
--- Zone: Escha_ZiTah (288)
+-- Zone: Escha - Zi'Tah (288)
 --
 -----------------------------------
 package.loaded["scripts/zones/Escha_ZiTah/TextIDs"] = nil;
 -----------------------------------
 
+require("scripts/zones/Escha_ZiTah/TextIDs");
 require("scripts/globals/settings");
 require("scripts/globals/zone");
-require("scripts/zones/Escha_ZiTah/TextIDs");
 
 -----------------------------------
 -- onInitialize
@@ -17,43 +17,47 @@ require("scripts/zones/Escha_ZiTah/TextIDs");
 function onInitialize(zone)
 end;
 
------------------------------------		
--- onZoneIn		
------------------------------------		
+-----------------------------------
+-- onZoneIn
+-----------------------------------
 
-function onZoneIn(player,prevZone)		
-	local cs = -1;	
-	return cs;	
-end;		
+function onZoneIn( player, prevZone)
+    local cs = -1;
 
------------------------------------		
--- onConquestUpdate		
------------------------------------		
+    if (player:getXPos() == 0 and player:getYPos() == 0 and player:getZPos() == 0) then
+        -- player:setPos(x, y, z, rot);
+    end
+
+    return cs;
+end;
+-----------------------------------
+-- onConquestUpdate
+-----------------------------------
 
 function onConquestUpdate(zone, updatetype)
 end;
 
------------------------------------	
--- onRegionEnter	
------------------------------------	
+-----------------------------------
+-- onRegionEnter
+-----------------------------------
 
-function onRegionEnter(player,region)	
-end;	
+function onRegionEnter( player, region)
+end;
 
------------------------------------	
--- onEventUpdate	
------------------------------------	
+-----------------------------------
+-- onEventUpdate
+-----------------------------------
 
-function onEventUpdate(player,csid,option)	
-	--printf("CSID: %u",csid);
-	--printf("RESULT: %u",option);
-end;	
+function onEventUpdate( player, csid, option)
+    -- printf("CSID: %u",csid);
+    -- printf("RESULT: %u",option);
+end;
 
------------------------------------		
--- onEventFinish		
------------------------------------		
+-----------------------------------
+-- onEventFinish
+-----------------------------------
 
-function onEventFinish(player,csid,option)		
-	--printf("CSID: %u",csid);	
-	--printf("RESULT: %u",option);	
-end;		
+function onEventFinish( player, csid, option)
+    -- printf("CSID: %u",csid);
+    -- printf("RESULT: %u",option);
+end;
